server/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
pydantic-settings>=2.6,<3
httpx>=0.28,<1
